requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-parler==2.3
djangorestframework==3.14.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
